Bf (legacy function)

The following R code is based on the Baguley and Kaye (2010) R code for the Dienes (2008) calculator, changing the likelihood function for the data from a Normal distribution to a t-distribution. This is a way of accounting for the variance of observations being unknown in advance of the data - thus no correction factor need be applied to the SE when this calculator is used. From http://www.lifesci.sussex.ac.uk/home/Zoltan_Dienes/inference/Bf

Bf(sd = 0.1, obtained = 0, dfdata = 99, uniform = 0, lower = 0,
  upper = 1, meanoftheory = 0, sdtheory = 1, tail = 2)

Arguments

sd
observed standard error

obtained
obtained mean

dfdata
observed degrees of freedom

uniform
is it uniform?

lower
lower-bound (if uniform)

upper
upper-bound (if uniform)

meanoftheory
theory mean

sdtheory
theory SD

tail
number of tails

Value

list